Update Lexer.py
There was an error on self.retreat being used that causes it to reach further negative numbers  than -1. If self.pos is -2, then a self.advance will result in self.pos -1 which is the end of the line. I made it so that self.retreat can't go lower than -1 in handling keywords turned identifiers.